fix: IssueService 생성
